mid-project-347359563/services/notification-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
aiosmtplib
aio_pika
cryptography
